backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pytest
flask-migrate
requests
